# lib/api-zod/src/generated/types/caseStudy.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaseStudy {
  id: number;
  slug: string;
  title: string;
  client: string;
  category: string;
  summary: string;
  problem: string;
  solution: string;
  result: string;
  metric: string;
  imageUrl: string;
  imageAlt: string;
  displayOrder: number;
}