# lib/api-zod/src/generated/types/caseStudyUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStudyInput } from './caseStudyInput';

export type CaseStudyUpdate = CaseStudyInput;