# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostInput } from './postInput';

export type PostUpdate = PostInput;