# lib/api-zod/src/generated/types/reelUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReelInput } from './reelInput';

export type ReelUpdate = ReelInput;